fix: delete transfer-encoding header to avoid 502
In Local 9.1 we updated Local's nginx router config to support
Server Sent Events, which involved changing from HTTP 1.0
to HTTP 1.1 to enable persistent connections:

https://github.com/getflywheel/flywheel-local/pull/2025/files

This accidentally broke Live Reload — Local's router would 502
if users had Live Reload active on a site, with an error in
Local's nginx router logs like this:

upstream sent "Content-Length" and "Transfer-Encoding" headers
at the same time

Under HTTP 1.0 this was fine, but HTTP 1.1 does not allow both headers.

This commit deletes the transfer-encoding header so that only
content-length remains.

Technically, it would be better to delete content-length than
transfer-encoding to meet the RFC 7230 Section 3.3.3 spec:
https://datatracker.ietf.org/doc/html/rfc7230#section-3.3.3

But attempting that here fails — content-length gets added
after the proxyRes callback, which seems like the last point
we're able to filter headers so we're not able to remove it here.